refactor-#164: move payment controller to controllers
