+++ src/microservices/payment/payment.module.ts
@@ -5,7 +5,7 @@ import { createClientAsyncOptions } from '@utils/client';
 import { BookingModule } from '@booking/booking.module';
 import { BookingService } from '@booking/services';
 import { PaymentService } from './services';
-import { PaymentController } from './rest/payment.controller';
+import { PaymentController } from './rest/controllers';
 import { PaymentMutationResolver } from './graphql/payment-mutation.resolver';
 import { PaymentQueryResolver } from './graphql/payment-query.resolver';
