+++ src/microservices/payment/rest/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './payment.controller';
