+++ src/microservices/payment/rest/controllers/payment.controller.ts
@@ -12,9 +12,9 @@ import {
 } from '@nestjs/common';
 
 import { AccessControlGuard } from '@auth/guards';
-import { PaymentService } from '../services';
-import { RestPayment } from './models/payment-rest';
-import { CreatePaymentDto } from './dto';
+import { PaymentService } from '../../services';
+import { RestPayment } from '../models';
+import { CreatePaymentDto } from '../dto';
 
 @Controller('payment')
 export class PaymentController {
