+++ src/microservices/payment/rest/models/index.ts
@@ -1,0 +1,1 @@
+export * from './payment-rest';
